Move json utils inside `autogpt/core`
