+++ autogpts/autogpt/autogpt/agent_factory/profile_generator.py
@@ -3,6 +3,7 @@ import logging
 
 from autogpt.config import AIDirectives, AIProfile, Config
 from autogpt.core.configuration import SystemConfiguration, UserConfigurable
+from autogpt.core.json_utils.utils import json_loads
 from autogpt.core.prompting import (
     ChatPrompt,
     LanguageModelClassification,
@@ -15,7 +16,6 @@ from autogpt.core.resource.model_providers.schema import (
     CompletionModelFunction,
 )
 from autogpt.core.utils.json_schema import JSONSchema
-from autogpt.json_utils.utilities import json_loads
 
 logger = logging.getLogger(__name__)
