+++ autogpts/autogpt/autogpt/core/json_utils/utils.py
@@ -1,5 +1,3 @@
-"""Utilities for the json_fixes package."""
-
 import io
 import logging
 import re
@@ -11,13 +9,14 @@ logger = logging.getLogger(__name__)
 
 
 def json_loads(json_str: str) -> Any:
-    """Parse a JSON string, this function is tolerant
-    to minor issues in the JSON string:
-    - Missing commas between elements
-    - Trailing commas or extra commas in objects
-    - Extraneous newlines and spaces outside of string literals
+    """Parse a JSON string, tolerating minor syntax issues:
+    - Missing, extra and trailing commas
+    - Extraneous newlines and whitespace outside of string literals
     - Inconsistent spacing after colons and commas
     - Missing closing brackets or braces
+    - Numbers: binary, hex, octal, trailing and prefixed decimal points
+    - Different encodings
+    - Surrounding markdown code block
     - Comments
 
     Args:
@@ -26,6 +25,13 @@ def json_loads(json_str: str) -> Any:
     Returns:
         The parsed JSON object, same as built-in json.loads.
     """
+    # Remove possible code block
+    pattern = r"```(?:json|JSON)*([\s\S]*?)```"
+    match = re.search(pattern, json_str)
+
+    if match:
+        json_str = match.group(1).strip()
+
     error_buffer = io.StringIO()
     json_result = demjson3.decode(
         json_str, return_errors=True, write_errors=error_buffer
@@ -40,49 +46,47 @@ def json_loads(json_str: str) -> Any:
     return json_result.object
 
 
-def extract_dict_from_response(response_content: str) -> dict[str, Any]:
+def extract_dict_from_json(json_str: str) -> dict[str, Any]:
     # Sometimes the response includes the JSON in a code block with ```
     pattern = r"```(?:json|JSON)*([\s\S]*?)```"
-    match = re.search(pattern, response_content)
+    match = re.search(pattern, json_str)
 
     if match:
-        response_content = match.group(1).strip()
+        json_str = match.group(1).strip()
     else:
         # The string may contain JSON.
         json_pattern = r"{[\s\S]*}"
-        match = re.search(json_pattern, response_content)
+        match = re.search(json_pattern, json_str)
 
         if match:
-            response_content = match.group()
+            json_str = match.group()
 
-    result = json_loads(response_content)
+    result = json_loads(json_str)
     if not isinstance(result, dict):
         raise ValueError(
-            f"Response '''{response_content}''' evaluated to "
-            f"non-dict value {repr(result)}"
+            f"Response '''{json_str}''' evaluated to " f"non-dict value {repr(result)}"
         )
     return result
 
 
-def extract_list_from_response(response_content: str) -> list[Any]:
+def extract_list_from_json(json_str: str) -> list[Any]:
     # Sometimes the response includes the JSON in a code block with ```
     pattern = r"```(?:json|JSON)*([\s\S]*?)```"
-    match = re.search(pattern, response_content)
+    match = re.search(pattern, json_str)
 
     if match:
-        response_content = match.group(1).strip()
+        json_str = match.group(1).strip()
     else:
         # The string may contain JSON.
         json_pattern = r"\[[\s\S]*\]"
-        match = re.search(json_pattern, response_content)
+        match = re.search(json_pattern, json_str)
 
         if match:
-            response_content = match.group()
+            json_str = match.group()
 
-    result = json_loads(response_content)
+    result = json_loads(json_str)
     if not isinstance(result, list):
         raise ValueError(
-            f"Response '''{response_content}''' evaluated to "
-            f"non-list value {repr(result)}"
+            f"Response '''{json_str}''' evaluated to " f"non-list value {repr(result)}"
         )
     return result
