+++ autogpts/autogpt/autogpt/core/resource/model_providers/openai.py
@@ -15,6 +15,7 @@ from openai.types.chat import ChatCompletion
 from pydantic import SecretStr
 
 from autogpt.core.configuration import Configurable, UserConfigurable
+from autogpt.core.json_utils.utils import json_loads
 from autogpt.core.resource.model_providers.schema import (
     AssistantChatMessage,
     AssistantToolCall,
@@ -38,7 +39,6 @@ from autogpt.core.resource.model_providers.schema import (
     ModelTokenizer,
 )
 from autogpt.core.utils.json_schema import JSONSchema
-from autogpt.json_utils.utilities import json_loads
 
 _T = TypeVar("_T")
 _P = ParamSpec("_P")
